Crate lcms2

See Little CMS full documentation for more in-depth information about LCMS functions.

The main types you need to use in this crate are Profile and Transform

Structs

AllowCache

CIECAM02
CIE CAM02

CIELab

CIEXYZ

DisallowCache

Flags
Flags for creating Transform. Can be OR-ed together with |.

Locale
Language code from ISO-639/2 and region code from ISO-3166.

MLU
This represents owned Multi Localized Unicode type. Most methods are implemented on MLURef.

MLURef
This is a borrwed Multi Localized Unicode type. It holds Unicode strings associated with Locale.

NamedColorInfo
Color in the palette

NamedColorList
Owned version of NamedColorListRef

NamedColorListRef
Palette of colors with names

Pipeline
This is an owned version of PipelineRef.

PipelineRef
Pipelines are a convenient way to model complex operations on image data.

Profile
An ICC color profile

Stage
This is an owned version of Stage.

StageRef
Stage functions

StagesIter

ThreadContext
Per-thread context for multi-threaded operation.

ToneCurve
Owned version of ToneCurveRef

ToneCurveRef
Tone curves are powerful constructs that can contain curves specified in diverse ways.

Transform
Conversion between two ICC profiles.

ViewingConditions

Enums

ColorSpaceSignature

Error

InfoType

Intent

PixelFormat
Format of pixel is defined by one cmsUInt32Number, using bit fields as follows

ProfileClassSignature

Tag
Value of a tag in an ICC profile

TagSignature

Traits

CIELabExt
Delta E

CIEXYZExt
Chromatic adaptation

CIExzYExt
White point

CacheFlag

ColorSpaceSignatureExt

Functions

version
LCMS version

white_point_from_temp
Temperature <-> Chromaticity (Black body) Color temperature is a characteristic of visible light that has important applications.

Type Definitions

LCMSResult
This is a regular Result type with LCMS-specific Error