Struct lcms2::MLURef

pub struct MLURef(_);

This is a borrwed Multi Localized Unicode type. It holds Unicode strings associated with Locale.

Methods

impl MLURef

pub fn set_text_ascii(&mut self, text: &str, locale: Locale) -> bool

Fills an ASCII (7 bit) entry for the given Language and country.

pub fn set_text(&mut self, text: &str, locale: Locale) -> bool

Fills a UNICODE wide char (16 bit) entry for the given Language and country.

pub fn text_ascii(&self, locale: Locale) -> LCMSResult<CString>

Gets an ASCII (7 bit) entry for the given Language and country.

pub fn text(&self, locale: Locale) -> LCMSResult<String>

Gets a Unicode entry for the given Language and country

pub fn tanslations(&self) -> Vec<Locale>

Obtains the translations stored in a given multilocalized unicode object.

pub fn tanslation(&self, locale: Locale) -> LCMSResult<Locale>

Obtains the translation rule for given multilocalized unicode object.