Struct lcms2::PipelineRef

pub struct PipelineRef(_);

Pipelines are a convenient way to model complex operations on image data.

Each pipeline may contain an arbitrary number of stages. Each stage performs a single operation. Pipelines may be optimized to be executed on a certain format (8 bits, for example) and can be saved as LUTs in ICC profiles.

Methods

impl PipelineRef

pub fn cat(&mut self, append: &PipelineRef) -> bool

Appends pipeline given as argument at the end of this pipeline. Channel count must match.

pub fn stage_count(&self) -> usize

pub fn first_stage(&self) -> Option<&StageRef>

pub fn last_stage(&self) -> Option<&StageRef>

Important traits for StagesIter<'a>
impl<'a> Iterator for StagesIter<'a> type Item = &'a StageRef;

pub fn stages(&self) -> StagesIter

pub fn set_8bit(&mut self, on: bool) -> bool

pub fn input_channels(&self) -> usize

pub fn output_channels(&self) -> usize

pub fn eval<Value: FloatOrU16>(&self, input: &[Value], output: &mut [Value])

pub unsafe fn eval_unchecked<Value: FloatOrU16>(
&self,
input: &[Value],
output: &mut [Value]
)