Struct lcms2::ToneCurve

pub struct ToneCurve(_);

Owned version of ToneCurveRef

Methods

impl ToneCurve

pub fn new(gamma: f64) -> Self

Simplified wrapper to new_parametric. Builds a parametric curve of type 1.

pub fn new_tabulated(values: &[u16]) -> Self

Builds a tone curve based on a table of 16-bit values. Tone curves built with this function are restricted to 0…1.0 domain.

pub fn new_tabulated_float(values: &[f32]) -> Self

Builds a tone curve based on a table of floating point values. Tone curves built with this function are not restricted to 0…1.0 domain.

pub fn new_parametric(curve_type: u16, params: &[f64]) -> LCMSResult<Self>

See Table 52 in LCMS documentation for descriptino of the types.

Exponential

CIE 122-1966

IEC 61966-3

IEC 61966-2.1 (sRGB)

See PDF

Identical to 5, unbounded.

See PDF

See PDF

(108) S-Shaped sigmoidal

Always use 10-parameter slice for plug-in types.