Struct lcms2::PipelineRef

pub struct PipelineRef(_);

Pipelines are a convenient way to model complex operations on image data.

Each pipeline may contain an arbitrary number of stages. Each stage performs a single operation. Pipelines may be optimized to be executed on a certain format (8 bits, for example) and can be saved as LUTs in ICC profiles.

Methods

impl PipelineRef

fn cat(&mut self, append: &PipelineRef) -> bool

Appends pipeline given as argument at the end of this pipeline. Channel count must match.

fn stage_count(&self) -> usize

fn first_stage(&self) -> Option<&StageRef>

fn last_stage(&self) -> Option<&StageRef>

fn stages(&self) -> StagesIter

fn set_8bit(&mut self, on: bool) -> bool

fn input_channels(&self) -> usize

fn output_channels(&self) -> usize

fn eval<Value: FloatOrU16>(&self, input: &[Value], output: &mut [Value])

unsafe fn eval_unchecked<Value: FloatOrU16>(
&self,
input: &[Value],
output: &mut [Value]
)