Struct lcms2::StageRef

pub struct StageRef(_);

Stage functions

Stages are single-step operations that can be chained to create pipelines. Actual stage types does include matrices, tone curves, Look-up interpolation and user-defined. There are functions to create new stage types and a plug-in type to allow stages to be saved in multi profile elements tag types. See the plug-in API for further details.

Methods

impl StageRef

fn input_channels(&self) -> usize

fn output_channels(&self) -> usize

fn stage_type(&self) -> StageSignature